# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/me/saved-listings": {
      "get": {
        "operationId": "SavedController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Saved Listings"
        ]
      }
    },
    "/me/saved-listings/{listingId}": {
      "post": {
        "operationId": "SavedController_save",
        "parameters": [
          {
            "name": "listingId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Saved Listings"
        ]
      },
      "delete": {
        "operationId": "SavedController_remove",
        "parameters": [
          {
            "name": "listingId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Saved Listings"
        ]
      }
    },
...
  },
  "info": {
    "title": "RentCity Backend API",
    "description": "REST API contract for RentCity web, mobile app, and phone-based web app clients.",
    "version": "0.2.4",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}